refactor: Create theme folder, update Pack Category styles
