+++ src/components/Dashboard/PackCategory/PackCategory.css
@@ -1,10 +1,3 @@
-.table-container {
-	width: 100%;
-	text-align: center;
-	margin: 15px 0px;
-	display: flex;
-}
-
 .ui.table tfoot {
 	background-color: #f0f0f0;
 	font-size: 0.9em;
