+++ src/shared/theme/tidyTheme.ts
@@ -1,0 +1,7 @@
+export type TidyTheme = {
+	lightGrey: '#f0f0f0';
+};
+
+export const tidyTheme: TidyTheme = {
+	lightGrey: '#f0f0f0',
+};
