+++ src/views/Layout/AppWithRoutes.tsx
@@ -3,7 +3,7 @@ import { Loader } from 'semantic-ui-react';
 import { userRoutes, guestRoutes } from './Routes.tsx';
 import { useGetAuthStatusQuery } from '../../queries/userQueries.ts';
 import { ThemeProvider } from 'styled-components';
-import { createTheme } from './themeUtils.ts';
+import { createTheme } from '../../shared/theme/themeUtils.ts';
 
 const AppWithRoutes = () => {
 	const { isLoading, data } = useGetAuthStatusQuery();
